Bulk commit of offline work on Hunt the Wumpus.
At this point the code builds all relevant data structures for the cave,
connecting the rooms with tunnels and populating with bats/pits/wumpus. User
interaction progresses through providing the user with instructions, cave
information and room information.

The code does not yet allow the user to take actions like moving or shooting.